+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/JibExtension.java
@@ -253,6 +253,12 @@ public class JibExtension {
     return container.getPorts();
   }
 
+  @Internal
+  @Optional
+  boolean getUseCurrentTimestamp() {
+    return container.getUseCurrentTimestamp();
+  }
+
   @Input
   @Optional
   boolean getUseOnlyProjectCache() {